+++ procodile-example/src/procodile_example/cli.py
@@ -2,12 +2,17 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-from procodile.cli import get_cli
+from procodile_example import __version__
+from procodile.cli import new_cli
 
 # The CLI with a basic set of commands.
 # The `cli` is a Typer application of type `typer.Typer()`,
 # so can use the instance to register your own commands.
-cli = get_cli("procodile_example.processes:registry")
+cli = new_cli(
+    registry="procodile_example.processes:registry",
+    name="procodile-example",
+    version=__version__,
+)
 
 
 __all__ = ["cli"]